+++ apx_core/src/url/mod.rs
@@ -1,2 +1,5 @@
+//! URL types
+
+pub mod canonical;
 pub mod common;
 pub mod hostname;