+++ apx_sdk/src/lib.rs
@@ -11,7 +11,6 @@ mod http_client;
 pub mod http_server;
 pub mod identifiers;
 pub mod jrd;
-pub mod url;
 pub mod utils;
 
 pub use apx_core as core;